skills/archimedes-omega/requirements.txt
numpy>=1.21.0
scipy>=1.8.0
matplotlib>=3.5.0
pandas>=1.4.0
pytest>=7.0.0
nostr>=0.0.2
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
strawberry-graphql>=0.235.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
